Add zemplate
This template attempts to move the build script away from the
sources to be built. It provides a `srcs.nix' which is where
the sources to be built are intended to go, whilst also allowing
the flake to provide its own sources.

It attempts to be backward compatible with olderstyle nix scripts
and with newer flakes.